requirements.txt
langchain>=0.1.14
langchain-google-genai>=0.0.7
chromadb>=0.4.22
sentence-transformers>=2.2.2
huggingface-hub>=0.21.4
python-dotenv>=1.0.1
pandas>=2.2.2
numpy>=1.26.4
tqdm>=4.66.2
pydantic>=1.10.14
typing-extensions>=4.11.0